+++ src/jsMain/kotlin/dev/inmo/jsuikit/utils/AnyNotNull.kt
@@ -1,0 +1,11 @@
+package dev.inmo.jsuikit.utils
+
+fun <T> Array<T>.anyNotNull(): Boolean {
+    for (item in this) {
+        if (item != null) {
+            return true
+        }
+    }
+
+    return false
+}