+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.0.30
+
+* Add `UIKitComment`
+* Including of `Comment` element
+* Upfill `UIKitSubNav`
+
 ## 0.0.29
 
 * Unfilling of `UIKitText`